requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.14.3
blessed==1.20.0
bs4==0.0.2
colorama==0.4.6
diff-match-patch==20241021
Django==5.1.6
django-import-export==4.3.7
django-picklefield==3.3
django-q2==1.7.6
et_xmlfile==2.0.0
iniconfig==2.0.0
jinxed==1.3.0
lxml==6.0.2
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
psycopg2==2.9.10
pytest==8.3.4
pytest-django==4.10.0
python-dateutil==2.9.0.post0
pytz==2025.1
redis==3.5.3
setuptools==79.0.1
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.3
tablib==3.8.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.15.0
tzdata==2025.1
wcwidth==0.2.13
XlsxWriter==3.2.3
